requirements.txt
flask
openai-whisper
libretranslate
tensorflow
transformers
